update: user authentication messages and add more constraints for password
